Harden native FFI safety, enable strict-casts, and add invariant/contract test suites
- Unify package repository URLs to https://github.com/sigurdm/scientific_dart and bundle third-party licenses (Google Highway, miniz, PocketFFT).
- Document decision not to pass externalSize to NativeFinalizer.attach per Dart VM team guidance and rely on NDArray.scope / dispose().
- Upgrade ScratchArena.reset out-of-bounds assertion to a runtime StateError and add value equality to ScratchMarker.
- Compile C++ hook sources with -fno-exceptions, replace std::vector in custom_indexing.cpp with NoThrowBuffer, make CRC32 table init thread-safe noexcept, and verify mz_crc32 in NPZ decompression.
- Check native return codes in tile and roll, validate Windows .def exports, and automatically fall back from fetch mode to source mode when prebuilt releases are unavailable.
- Add NDARRAY_SANITIZE support in hook/build.dart for ASan/UBSan builds.
- Enable strict-casts: true across package:ndarray/lib.
- Add AST/FFI static invariant tests (test/meta/codebase_invariants_test.dart) and table-driven operational contract tests (test/meta/operation_contracts_test.dart).

+++ pkgs/ndarray/lib/src/operations/indexing.dart
@@ -871,7 +871,7 @@ NDArray<T> choose<T extends DTypeTag>(
       while (iter.moveNext()) {
         final coords = iter.coords;
         _mapCoordInPlace(coords, a.shape, aCoord);
-        var idxVal = a.getCell(aCoord);
+        var idxVal = a.getCell(aCoord) as int;
         if (isUint64 && idxVal < 0) {
           idxVal = mode == ChooseMode.wrap
               ? ((idxVal & 0x7FFFFFFFFFFFFFFF) % nChoices + twoPow63Mod) %
